Add PostgresSQL database migrations and configurations
Two database migration files have been created to set up the `users` and the `tasks` tables in a PostgreSQL database. This commit also introduces a configuration for establishing a connection to the PostgreSQL database. Helper tools have been added to convert environment variables to required datatypes and new utility functions for password hashing and validation are included as well.